Bug fixed for swagger documentation, separated class from Users refs socialappslab/appcivist-platform#239 Class not found in Swagger API documentation generation
