+++ app/views/password_change.scala.html
@@ -1,4 +1,4 @@
-@(changeForm: Form[controllers.Users.PasswordChange])
+@(changeForm: Form[controllers.PasswordChange])
 
 @import helper._
 @implicitFieldConstructor = @{ FieldConstructor(twitterBootstrapBasic.f) }
